+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v0.7.1]
+
+### Fixed
+- Skills & saving throw dropdown behaviour
+- Unused advance share setting removed from code
+
+### Added
+- Breakdown panel redesign
+- HP Breakdown toggle to settings
+- Rolled HP arrow indicator showing deviation from average
+
 ## [v0.7.0]
 
 ### Fixed